Move set_request_state to tasks.utils
CLOUDBLD-4727

Most of the methods than communicate with the Cachito API are now in
tasks.utils, set_request_state should be there too.

Also, set_request_state is no longer an `@app.task`, it is never used as
a task.
